# claims-app/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Claims Management API",
    "description": "API for managing customer claims and complaints",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "Claims",
      "description": "Claims management API"
    },
    {
      "name": "Hello",
      "description": "Test endpoint for API verification"
    }
  ],
  "paths": {
    "/api/claims": {
      "get": {
        "tags": ["Claims"],
        "summary": "Get all claims",
        "description": "Retrieves all claims with optional filtering by status and text search",
        "operationId": "getAllClaims",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by claim status (optional)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "Ingresado",
                "En Proceso",
                "Resuelto",
                "Cerrado",
                "Rechazado",
                "Escalado",
                "Pendiente Información"
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search text in title, description or code (optional)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Claims retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClaimResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Claims"],
        "summary": "Create a new claim",
        "description": "Creates a new claim with the provided details",
        "operationId": "createClaim",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Claim created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/claims/{id}/status": {
      "post": {
        "tags": ["Claims"],
        "summary": "Update claim status",
        "description": "Adds a new status update to the specified claim",
        "operationId": "updateClaimStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the claim to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Status updated successfully"
          },
          "400": {
            "description": "Invalid status data"
          },
          "404": {
            "description": "Claim not found"
          }
        }
      }
    },
    "/api/claims/{id}/attachments": {
      "post": {
        "tags": ["Claims"],
        "summary": "Upload attachment",
        "description": "Uploads a file attachment to the specified claim",
        "operationId": "uploadAttachment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the claim to add attachment to",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": ["file"],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "description": "File to upload",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Attachment uploaded successfully"
          },
          "400": {
            "description": "Invalid file or file is empty"
          },
          "404": {
            "description": "Claim not found"
          }
        }
      }
    },
    "/hello": {
      "get": {
        "tags": ["Hello"],
        "summary": "Test endpoint",
        "description": "Simple test endpoint to verify API is running correctly",
        "operationId": "hello",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/claims/{id}": {
      "get": {
        "tags": ["Claims"],
        "summary": "Get claim details",
        "description": "Retrieves detailed information about a specific claim",
        "operationId": "getClaimById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the claim to retrieve",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Claim details retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Claim not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/claims/export/pdf": {
      "get": {
        "tags": ["Claims"],
        "summary": "Export claims to PDF",
        "description": "Exports all claims to a PDF file and returns it as base64 encoded content",
        "operationId": "exportClaimsToPdf",
        "responses": {
          "200": {
            "description": "PDF generated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PdfExportResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error generating PDF",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PdfExportResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateClaimRequest": {
        "required": ["customerId", "description", "title"],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "customerId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ClaimResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "customerId": {
            "type": "integer",
            "format": "int64"
          },
          "currentStatus": {
            "type": "string",
            "enum": [
              "Ingresado",
              "En Proceso",
              "Resuelto",
              "Cerrado",
              "Rechazado",
              "Escalado",
              "Pendiente Información"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ClaimStatusRequest": {
        "required": ["status"],
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "New status for the claim",
            "enum": [
              "Ingresado",
              "En Proceso",
              "Resuelto",
              "Cerrado",
              "Rechazado",
              "Escalado",
              "Pendiente Información"
            ]
          },
          "notes": {
            "type": "string",
            "description": "Optional notes for the status change"
          },
          "asesor_email": {
            "type": "string",
            "description": "Email of the advisor handling the claim"
          }
        },
        "description": "Request to update claim status"
      },
      "ClaimAttachment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fileName": {
            "type": "string"
          },
          "fileType": {
            "type": "string"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ClaimDetailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "customerId": {
            "type": "integer",
            "format": "int64"
          },
          "currentStatus": {
            "type": "string",
            "enum": [
              "Ingresado",
              "En Proceso",
              "Resuelto",
              "Cerrado",
              "Rechazado",
              "Escalado",
              "Pendiente Información"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time"
          },
          "statusHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClaimStatusHistory"
            }
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClaimAttachment"
            }
          }
        }
      },
      "ClaimStatusHistory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "Ingresado",
              "En Proceso",
              "Resuelto",
              "Cerrado",
              "Rechazado",
              "Escalado",
              "Pendiente Información"
            ]
          },
          "notes": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PdfExportResponse": {
        "required": ["filename", "pdfContent", "totalClaims"],
        "type": "object",
        "properties": {
          "pdfContent": {
            "type": "string",
            "description": "Base64 encoded PDF content"
          },
          "filename": {
            "type": "string",
            "description": "Filename for the PDF"
          },
          "totalClaims": {
            "type": "integer",
            "description": "Total number of claims exported",
            "format": "int32"
          }
        },
        "description": "Response containing the exported PDF as base64"
      }
    }
  }
}